requirements.txt
asttokens==2.2.1
backcall==0.2.0
catboost==1.1.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
dnspython==2.3.0
executing==1.2.0
fonttools==4.38.0
graphviz==0.20.1
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
joblib==1.2.0
jupyter-client==8.0.3
jupyter-core==5.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.0
prompt-toolkit==3.0.37
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
tenacity==8.2.1
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
xgboost==1.7.4
zipp==3.14.0
